Skip updating initial lines that have not changed
In many cases, the most frequently-changing part of the logUpdate output
is at the bottom, as that is where users expect to see output.  This
patch optimizes for that case by skipping updating any initial lines
that have not changed.  For output spanning many lines, this can help
avoid unsightly flicker on update.